requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.1.7
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.1
idna==2.10
jwcrypto==0.8
oauth2client==4.1.3
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.4.3
python-jwt==3.3.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
